Fixed issue where routes cannot be found with Polyglot is used with Laravel 5.
Laravel 5 instantiates the router really early in the app lifecycle. A
router will be created and injected when the kernel is instantiated.
This means it's done before any service providers are registered or
booted.

Attempting to create a new shared instance to supercede the old router
is problematic. The kernel does not attempt to fetch a potentially
different router after all services are booted. It simply dispatches
the request with the router it was injected with.

To integrate with Laravel 5 properly, the router binding needs to
be declared in bootstrap/app.php, before the kernel is instantiated.
Other attempts to replace the router is not needed.